+++ sales-copilot/supabase/functions/hubspot-auto-training/index.ts
@@ -19,93 +19,9 @@ declare const Deno: {
   };
 };
 
-const logger = new Logger('hubspot-auto-training');
+const logger = new Logger('hubspot-auto-training-v2');
 
-interface DealCriteria {
-  dealstage: string;
-  amount?: number;
-  closedate?: string;
-  createdate: string;
-}
-
-interface DealMetadata {
-  deal_id: string;
-  deal_value: number;
-  conversion_days: number;
-  pipeline: string;
-  dealstage: string;
-  won_date?: string;
-  lost_date?: string;
-  related_contacts: string[];
-  related_companies: string[];
-  days_in_pipeline: number;
-  stage_changes: number;
-}
-
-interface ContactMetadata {
-  contact_id: string;
-  lifecycle_stage: string;
-  lead_source: string;
-  total_deals_participated: number;
-  total_won_deals: number;
-  total_lost_deals: number;
-  first_conversion_date?: string;
-  last_activity_date?: string;
-  engagement_score: number;
-  deal_participation_rate: number;
-  industry?: string;
-  company_size?: string;
-}
-
-interface CompanyMetadata {
-  company_id: string;
-  industry?: string;
-  company_size?: string;
-  total_revenue?: number;
-  country?: string;
-  city?: string;
-  total_deals: number;
-  won_deals: number;
-  lost_deals: number;
-  average_deal_size: number;
-  average_sales_cycle: number;
-  engagement_score: number;
-  first_deal_date?: string;
-  last_deal_date?: string;
-}
-
-interface SearchFilter {
-  propertyName: string;
-  operator: string;
-  value?: string;
-  values?: string[];
-}
-
-interface SearchSort {
-  propertyName: string;
-  direction: 'ASCENDING' | 'DESCENDING';
-}
-
-interface SearchRequest {
-  filterGroups: Array<{
-    filters: SearchFilter[];
-  }>;
-  properties?: string[];
-  limit?: number;
-  after?: string;
-  sorts?: SearchSort[];
-}
-
-function calculateConversionDays(properties: DealCriteria): number {
-  const createDate = new Date(properties.createdate);
-  const closeDate = properties.closedate ? new Date(properties.closedate) : null;
-  
-  if (!closeDate) return 0;
-  
-  return Math.ceil((closeDate.getTime() - createDate.getTime()) / (1000 * 60 * 60 * 24));
-}
-
-// Add these helper functions for rate limiting and batching
+// Helper functions
 async function sleep(ms: number) {
   return new Promise(resolve => setTimeout(resolve, ms));
 }
@@ -118,22 +34,187 @@ function chunkArray<T>(array: T[], size: number): T[][] {
   return chunks;
 }
 
+function calculateConversionDays(properties: any): number {
+  const createDate = new Date(properties.createdate);
+  const closeDate = properties.closedate ? new Date(properties.closedate) : null;
+  
+  if (!closeDate) return 0;
+  
+  return Math.ceil((closeDate.getTime() - createDate.getTime()) / (1000 * 60 * 60 * 24));
+}
+
+/**
+ * Handles API calls with automatic token refresh
+ */
+async function handleApiCall<T>(
+  hubspotClient: HubspotClient,
+  portalId: string | number,
+  refreshToken: string,
+  apiCall: () => Promise<T>
+): Promise<T> {
+  try {
+    return await apiCall();
+  } catch (error) {
+    logger.error(`API call error for portal ${portalId}:`, { message: error.message, status: error.status });
+
+    // Check if the error is due to token expiration
+    if (error.status === 401) {
+      logger.info(`Refreshing expired token for portal ${portalId}`);
+      
+      try {
+        // Refresh the token
+        const supabase = createClient(
+          Deno.env.get('SUPABASE_URL')!,
+          Deno.env.get('SUPABASE_SERVICE_ROLE_KEY')!
+        );
+
+        const refreshResult = await hubspotClient.refreshToken(
+          refreshToken,
+          Deno.env.get('HUBSPOT_CLIENT_ID')!,
+          Deno.env.get('HUBSPOT_CLIENT_SECRET')!
+        );
+
+        if (!refreshResult || !refreshResult.access_token) {
+          throw new Error('Failed to refresh token: Invalid response');
+        }
+
+        // Update the client with the new token
+        hubspotClient.setToken(refreshResult.access_token);
+
+        // Encrypt the new tokens and update in the database
+        const encryptedAccessToken = await encrypt(
+          refreshResult.access_token,
+          Deno.env.get('ENCRYPTION_KEY')!
+        );
+
+        const encryptedRefreshToken = await encrypt(
+          refreshResult.refresh_token,
+          Deno.env.get('ENCRYPTION_KEY')!
+        );
+
+        const { error: updateError } = await supabase
+          .from('hubspot_accounts')
+          .update({
+            access_token: encryptedAccessToken,
+            refresh_token: encryptedRefreshToken,
+            expires_at: new Date(Date.now() + 1800 * 1000).toISOString(), // Default to 30 minutes
+            updated_at: new Date().toISOString()
+          })
+          .eq('portal_id', portalId);
+
+        if (updateError) {
+          throw new Error(`Failed to update tokens in database: ${updateError.message}`);
+        }
+
+        logger.info(`Successfully refreshed token for portal ${portalId}`);
+
+        // Retry the API call with the new token
+        return await apiCall();
+      } catch (refreshError) {
+        logger.error(`Token refresh failed for portal ${portalId}:`, refreshError);
+        throw new Error(`Token refresh failed: ${refreshError.message}`);
+      }
+    }
+
+    // For other errors, propagate them
+    throw error;
+  }
+}
+
+// Enhance filterVectorsToUpdate to check for existing vectors
+function filterVectorsToUpdate(
+  dealsToProcess: any[], 
+  existingVectors: any[]
+): any[] {
+  try {
+    // Create a map of existing vectors by ID for faster lookup
+    const existingVectorsMap = new Map<string, any>();
+    
+    if (Array.isArray(existingVectors) && existingVectors.length > 0) {
+      existingVectors.forEach(vector => {
+        if (vector && vector.id) {
+          existingVectorsMap.set(vector.id.toString(), vector);
+          logger.info(`Adding vector ${vector.id} to existing vectors map`);
+        }
+      });
+      
+      logger.info(`Created existing vectors map with ${existingVectorsMap.size} entries`);
+    } else {
+      logger.info('No existing vectors found, will process all deals');
+      return dealsToProcess;
+    }
+    
+    // Filter deals to only include those that need updating
+    const dealsToUpdate = dealsToProcess.filter(deal => {
+      const dealId = deal.id.toString();
+      
+      // Check if vector exists in existing vectors
+      const existingVector = existingVectorsMap.get(dealId);
+      
+      if (!existingVector) {
+        logger.info(`Deal ${dealId} not found in existing vectors, will process`);
+        return true;
+      }
+      
+      logger.info(`Deal ${dealId} already exists in Pinecone, skipping update`);
+      return false;
+    });
+    
+    logger.info(`Filtered deals to update: ${dealsToUpdate.length} out of ${dealsToProcess.length}`);
+    return dealsToUpdate;
+  } catch (error) {
+    logger.error(`Error filtering vectors to update: ${error.message}`);
+    // If filtering fails, return all deals to be safe
+    return dealsToProcess;
+  }
+}
+
+/**
+ * Fetches deals for training from HubSpot
+ */
 async function getDealsForTraining(
   hubspotClient: HubspotClient, 
   type: 'ideal' | 'nonideal',
   portalId: string,
-  refreshToken: string
-): Promise<any[]> {
+  refreshToken: string,
+  documentPackager: DocumentPackager,
+  openai: any,
+  pineconeClient: PineconeClient
+): Promise<{ total: number, processed: number }> {
   try {
+    const dealStage = type === 'ideal' ? 'closedwon' : 'closedlost';
+    
+    // Define time range - last 90 days
     const ninety_days_ago = new Date();
     ninety_days_ago.setDate(ninety_days_ago.getDate() - 90);
-    const today = new Date();
-    const allDeals: any[] = [];
+    
+    // Define propertyNames to fetch
+    const propertyNames = [
+      'dealname',
+      'createdate',
+      'closedate',
+      'amount',
+      'dealstage',
+      'hs_lastmodifieddate',
+      'hs_object_id',
+      'description',
+      'notes_last_updated'
+    ];
+    
+    let totalDeals = 0;
+    let processedDeals = 0;
+    const namespace = `hubspot-${portalId}`;
+    
+    logger.info(`Fetching ${dealStage} deals using namespace ${namespace}`);
+    
     let hasMore = true;
     let after: string | null = null;
 
+    logger.info(`Fetching ${type} deals using namespace ${namespace}`);
+
+    // Pagination loop to fetch deals in batches
     while (hasMore) {
-      await sleep(2000);
+      await sleep(2000); // Rate limiting
 
       const searchCriteria = {
         filterGroups: [{
@@ -146,7 +227,7 @@ async function getDealsForTraining(
             {
               propertyName: 'dealstage',
               operator: 'EQ',
-              value: type === 'ideal' ? 'closedwon' : 'closedlost'
+              value: dealStage
             }
           ]
         }],
@@ -156,754 +237,580 @@ async function getDealsForTraining(
             direction: 'DESCENDING'
           }
         ],
-        properties: [
-          'dealname',
-          'amount',
-          'closedate',
-          'createdate',
-          'dealstage',
-          'pipeline',
-          'hs_lastmodifieddate',
-          'hs_date_entered_closedwon',
-          'hs_date_entered_closedlost'
-        ],
-        limit: 25,
+        properties: propertyNames,
+        limit: 10, // Fetch 10 at a time (reduced from 25 for more manageable batches)
         associations: ['contacts', 'companies'],
         ...(after ? { after } : {})
       };
-
-      logger.info(`Searching for ${type} deals with criteria:`, JSON.stringify(searchCriteria, null, 2));
       
-      const deals = await handleApiCall(
+      // Make API call with token refresh handling
+      const dealsResponse = await handleApiCall(
         hubspotClient,
         portalId,
         refreshToken,
         () => hubspotClient.searchRecords('deals', searchCriteria)
       );
       
-      logger.info(`Found ${deals.total || 0} total ${type} deals, fetched ${deals.results?.length || 0} in this batch`);
+      logger.info(`Found ${dealsResponse.total || 0} total ${type} deals, fetched ${dealsResponse.results?.length || 0} in this batch`);
       
-      if (deals.results?.length) {
-        // Log some sample data to verify associations
-        const sampleDeal = deals.results[0];
-        logger.info('Sample deal associations:', {
-          dealId: sampleDeal.id,
-          hasAssociations: !!sampleDeal.associations,
-          contactsCount: sampleDeal.associations?.contacts?.results?.length || 0,
-          companiesCount: sampleDeal.associations?.companies?.results?.length || 0
-        });
+      if (dealsResponse.results?.length) {
+        totalDeals += dealsResponse.results.length;
         
-        allDeals.push(...deals.results);
+        // Get the batch of deals to process
+        let dealBatch = dealsResponse.results;
+        
+        // Get the deal IDs from this batch
+        const batchDealIds = dealBatch.map(deal => deal.id.toString());
+        logger.info(`Fetched ${batchDealIds.length} deals in this batch for ${namespace}, checking which ones need updating`);
+        
+        // Fetch existing vectors for just these deals using the correct namespace
+        try {
+          const existingVectors = await getExistingVectors(pineconeClient, batchDealIds, namespace);
+          
+          if (existingVectors.length > 0) {
+            logger.info(`Found ${existingVectors.length} existing vectors in namespace ${namespace}, filtering deals`);
+            
+            // Filter to only deals that need updating using our enhanced function
+            dealBatch = filterVectorsToUpdate(dealBatch, existingVectors);
+            logger.info(`After filtering: ${dealBatch.length} out of ${dealsResponse.results.length} deals need updating`);
+          } else {
+            logger.info(`No existing vectors found in namespace ${namespace}, will process all deals`);
+          }
+        } catch (filterError) {
+          logger.error(`Error filtering vectors to update for ${namespace}: ${filterError.message}`);
+          // Continue with all deals if filtering fails
+        }
+        
+        // Skip further processing if no deals need updating
+        if (dealBatch.length === 0) {
+          logger.info(`All deals in this batch already exist in Pinecone, skipping processing`);
+          continue;
+        }
+        
+        // Process each deal in this filtered batch individually
+        for (const deal of dealBatch) {
+          try {
+            // Defensive checks
+            if (!deal || !deal.id) {
+              logger.error('Invalid deal object', { deal });
+              continue;
+            }
+
+            logger.info(`Processing deal ${deal.id}, ${deal.properties?.dealname || 'unnamed'}`);
+
+            try {
+              // Get full details for the deal
+              const fullDeal = await handleApiCall(
+                hubspotClient,
+                portalId,
+                refreshToken,
+                () => hubspotClient.getRecord('deals', deal.id, [
+                  'dealname',
+                  'amount',
+                  'closedate',
+                  'createdate',
+                  'dealstage',
+                  'pipeline',
+                  'hs_lastmodifieddate',
+                  'hs_date_entered_closedwon',
+                  'hs_date_entered_closedlost',
+                  'hs_deal_stage_probability',
+                  'hs_pipeline_stage',
+                  'hs_time_in_pipeline',
+                  'hs_time_in_dealstage',
+                  'hs_deal_stage_changes'
+                ])
+              );
+
+              if (fullDeal) {
+                // Preserve the associations from the search results
+                fullDeal.associations = deal.associations;
+                
+                // Process this single deal right away
+                await processSingleDeal(
+                  fullDeal,
+                  type,
+                  hubspotClient,
+                  documentPackager,
+                  openai,
+                  pineconeClient, 
+                  portalId,
+                  namespace
+                );
+                processedDeals++;
+                logger.info(`Successfully processed deal ${fullDeal.id} (${processedDeals}/${totalDeals})`);
+              }
+            } catch (error) {
+              logger.error(`Error processing deal ${deal.id}:`, error);
+            }
+            
+            await sleep(3000); // Rate limiting between deals
+          } catch (error) {
+            logger.error(`Error handling deal ${deal?.id || 'unknown'}:`, error);
+          }
+        }
+
+        logger.info(`Completed batch of ${dealBatch.length} deals (total: ${processedDeals}/${totalDeals})`);
+        await sleep(5000); // Rate limiting between batches
       }
 
-      hasMore = deals.paging?.next?.after !== undefined;
-      after = deals.paging?.next?.after || null;
+      // Check if more pages are available
+      hasMore = dealsResponse.paging?.next?.after !== undefined;
+      after = dealsResponse.paging?.next?.after || null;
 
       if (hasMore) {
         logger.info(`More deals available, continuing with after=${after}`);
-        await sleep(5000);
       }
     }
 
-    logger.info(`Total ${type} deals collected: ${allDeals.length}`);
-    
-    // Process full deal details in smaller batches
-    const fullDeals: any[] = [];
-    const dealBatches = chunkArray(allDeals, 5);
-
-    for (const batch of dealBatches) {
-      logger.info(`Processing batch of ${batch.length} deals`);
-      
-      for (const deal of batch) {
-        try {
-          // Log the raw deal object to see what we're working with
-          logger.info('Raw deal object:', {
-            id: deal?.id,
-            properties: deal?.properties ? Object.keys(deal.properties) : 'no properties',
-            associations: deal?.associations ? Object.keys(deal.associations) : 'no associations'
-          });
-
-          // Add defensive checks
-          if (!deal) {
-            logger.error('Deal object is null or undefined');
-            continue;
-          }
-
-          if (!deal.id) {
-            logger.error('Deal has no ID');
-            continue;
-          }
-
-          await sleep(3000);
-          logger.info(`Starting to fetch full deal details for ${deal.id}`);
-
-          try {
-            const fullDeal = await handleApiCall(
-              hubspotClient,
-              portalId,
-              refreshToken,
-              () => hubspotClient.getRecord('deals', deal.id, [
-                'dealname',
-                'amount',
-                'closedate',
-                'createdate',
-                'dealstage',
-                'pipeline',
-                'hs_lastmodifieddate',
-                'hs_date_entered_closedwon',
-                'hs_date_entered_closedlost',
-                'hs_deal_stage_probability',
-                'hs_pipeline_stage',
-                'hs_time_in_pipeline',
-                'hs_time_in_dealstage',
-                'hs_deal_stage_changes'
-              ])
-            );
-
-            logger.info(`Successfully fetched full deal details for ${deal.id}`);
-            
-            if (fullDeal) {
-              // Log the full deal object
-              logger.info('Full deal details:', {
-                id: fullDeal.id,
-                properties: fullDeal.properties ? Object.keys(fullDeal.properties) : 'no properties',
-                hasAssociations: !!deal.associations
-              });
-
-              // Preserve the associations from the search results
-              fullDeal.associations = deal.associations;
-              fullDeals.push(fullDeal);
-              logger.info(`Added deal ${deal.id} to fullDeals array`);
-              await sleep(1000);
-            } else {
-              logger.error(`getRecord returned null for deal ${deal.id}`);
-            }
-          } catch (getRecordError) {
-            logger.error(`Error in getRecord for deal ${deal.id}:`, {
-              error: getRecordError.message,
-              status: getRecordError.status,
-              response: getRecordError.response,
-              stack: getRecordError.stack
-            });
-            await sleep(5000);
-          }
-        } catch (error) {
-          logger.error(`Error in main deal processing loop for deal ${deal?.id || 'unknown'}:`, {
-            error: error.message,
-            stack: error.stack,
-            dealObject: deal ? 'exists' : 'null'
-          });
-          await sleep(5000);
-        }
-      }
-
-      logger.info(`Batch processing complete. Processed ${batch.length} deals, got ${fullDeals.length} full deals`);
-      await sleep(8000);
-    }
-
-    return fullDeals;
+    logger.info(`Completed processing ${type} deals: ${processedDeals}/${totalDeals}`);
+    return { total: totalDeals, processed: processedDeals };
   } catch (error) {
-    logger.error(`Error fetching ${type} deals:`, error);
+    logger.error(`Error fetching and processing ${type} deals:`, error);
     throw error;
   }
 }
 
+/**
+ * Process a single deal including its associations
+ */
+async function processSingleDeal(
+  deal: any,
+  classification: string,
+  hubspotClient: any,
+  documentPackager: DocumentPackager,
+  openai: any,
+  pineconeClient: PineconeClient,
+  portalId: string,
+  namespace: string
+): Promise<void> {
+  const processingStart = Date.now();
+  
+  logger.info(`Processing deal ${deal.id} (${classification}) in namespace ${namespace}`);
+  
+  try {
+    // Get all contacts and companies associated with this deal
+    // Always fetch fresh associations to ensure we're getting all related records
+    try {
+      const associationsResult = await hubspotClient.getAssociations(deal.id, 'deal');
+      if (associationsResult?.results) {
+        // Extract contact and company IDs from associations
+        const contactIds = associationsResult.results.contacts?.map(a => a.id) || [];
+        const companyIds = associationsResult.results.companies?.map(a => a.id) || [];
+        
+        const totalAssociations = contactIds.length + companyIds.length;
+        logger.info(`Found ${totalAssociations} associations for deal ${deal.id}`);
+        
+        // Create or update the associations structure
+        if (!deal.associations) deal.associations = {};
+        if (contactIds.length > 0) {
+          deal.associations.contacts = {
+            results: contactIds.map(id => ({ id }))
+          };
+        }
+        if (companyIds.length > 0) {
+          deal.associations.companies = {
+            results: companyIds.map(id => ({ id }))
+          };
+        }
+      }
+    } catch (associationsError) {
+      logger.error(`Error fetching associations: ${associationsError.message}`);
+    }
+    
+    const associatedRecords = await getAssociatedRecords(hubspotClient, deal);
+    
+    // Process deal records
+    logger.info(`Creating documents...`);
+    const dealDocuments = [];
+    
+    try {
+      const dealDocs = await documentPackager.packageDocument(deal, 'deal', portalId);
+      if (dealDocs) {
+        dealDocuments.push(dealDocs);
+      }
+    } catch (dealError) {
+      logger.error(`Error creating deal document: ${dealError.message}`);
+    }
+    
+    // Process contact records
+    const contactDocuments = [];
+    
+    try {
+      for (const contact of associatedRecords.contacts) {
+        try {
+          const contactDoc = await documentPackager.packageDocument(contact, 'contact', portalId);
+          if (contactDoc) {
+            contactDocuments.push(contactDoc);
+          }
+        } catch (contactError) {
+          logger.error(`Error packaging contact ${contact.id}: ${contactError.message}`);
+        }
+      }
+    } catch (contactsError) {
+      logger.error(`Error creating contact documents: ${contactsError.message}`);
+    }
+    
+    // Process company records
+    const companyDocuments = [];
+    
+    try {
+      for (const company of associatedRecords.companies) {
+        try {
+          const companyDoc = await documentPackager.packageDocument(company, 'company', portalId);
+          if (companyDoc) {
+            companyDocuments.push(companyDoc);
+          }
+        } catch (companyError) {
+          logger.error(`Error packaging company ${company.id}: ${companyError.message}`);
+        }
+      }
+    } catch (companiesError) {
+      logger.error(`Error creating company documents: ${companiesError.message}`);
+    }
+    
+    // Calculate deal metadata for all records
+    const dealMetadata = {
+      deal_id: deal.id,
+      deal_value: parseFloat(deal.properties?.amount) || 0,
+      conversion_days: calculateConversionDays(deal.properties || {}),
+      pipeline: deal.properties?.pipeline || 'unknown',
+      dealstage: deal.properties?.dealstage || 'unknown',
+      days_in_pipeline: parseInt(deal.properties?.hs_time_in_pipeline) || 0,
+      classification
+    };
+    
+    // Combine all documents and add metadata
+    const allDocuments = [
+      ...dealDocuments.map(doc => ({
+        ...doc,
+        metadata: { ...doc.metadata, ...dealMetadata, record_type: 'deal' }
+      })),
+      ...contactDocuments.map(doc => ({
+        ...doc,
+        metadata: { ...doc.metadata, ...dealMetadata, record_type: 'contact' }
+      })),
+      ...companyDocuments.map(doc => ({
+        ...doc,
+        metadata: { ...doc.metadata, ...dealMetadata, record_type: 'company' }
+      }))
+    ];
+    
+    logger.info(`Created ${allDocuments.length} documents for deal ${deal.id}`);
+    
+    if (allDocuments.length === 0) {
+      logger.warn(`No documents created for deal ${deal.id}. Skipping embeddings.`);
+      return;
+    }
+    
+    // Generate embeddings for all documents
+    logger.info(`Generating embeddings for ${allDocuments.length} documents`);
+    let documentsWithEmbeddings = [];
+    
+    try {
+      // Process in batches to avoid rate limits
+      const batchSize = 10;
+      for (let i = 0; i < allDocuments.length; i += batchSize) {
+        const batch = allDocuments.slice(i, i + batchSize);
+        
+        const embeddingResponse = await openai.embeddings.create({
+          model: "text-embedding-3-large",
+          input: batch.map(doc => doc.content),
+          encoding_format: "float"
+        });
+        
+        const batchWithEmbeddings = batch.map((doc, index) => ({
+          ...doc,
+          embedding: embeddingResponse.data[index].embedding
+        }));
+        
+        documentsWithEmbeddings.push(...batchWithEmbeddings);
+        
+        if (i + batchSize < allDocuments.length) {
+          await sleep(500);
+        }
+      }
+    } catch (embeddingsError) {
+      logger.error(`Error generating embeddings: ${embeddingsError.message}`);
+      throw embeddingsError;
+    }
+    
+    if (documentsWithEmbeddings.length === 0) {
+      logger.warn(`No documents with embeddings created for deal ${deal.id}.`);
+      return;
+    }
+    
+    // Upsert to Pinecone
+    logger.info(`Upserting ${documentsWithEmbeddings.length} vectors to namespace ${namespace}`);
+    try {
+      // Create vectors
+      const vectors = documentsWithEmbeddings.map(doc => ({
+        id: doc.metadata.id.toString(),
+        values: Array.from(doc.embedding),
+        metadata: {
+          ...doc.metadata
+        }
+      }));
+      
+      // Before upserting, check if these vectors already exist with the same deal metadata
+      const vectorIds = vectors.map(v => v.id);
+      try {
+        // Fetch existing vectors in batches (limit by Pinecone API)
+        const batchSize = 100;
+        const existingVectors = new Map();
+        
+        for (let i = 0; i < vectorIds.length; i += batchSize) {
+          const batchIds = vectorIds.slice(i, i + batchSize);
+          
+          try {
+            const fetchResult = await pineconeClient.query(namespace, null, { id: { $in: batchIds } }, batchIds.length);
+            
+            if (fetchResult.matches?.length > 0) {
+              // Store existing vectors to compare later
+              fetchResult.matches.forEach(vector => {
+                existingVectors.set(vector.id, {
+                  metadata: vector.metadata
+                });
+              });
+            }
+          } catch (fetchBatchError) {
+            logger.error(`Error fetching batch of vectors: ${fetchBatchError.message}`);
+            // Continue with next batch
+          }
+        }
+        
+        // Filter vectors that have changed
+        const vectorsToUpsert = vectors.filter(vector => {
+          // If vector doesn't exist, include it
+          if (!existingVectors.has(vector.id)) {
+            return true;
+          }
+          
+          // If vector exists, check if metadata has changed
+          const existingVector = existingVectors.get(vector.id);
+          // Compare deal metadata
+          const dealMetadataChanged = 
+            existingVector.metadata.deal_id !== vector.metadata.deal_id ||
+            existingVector.metadata.deal_value !== vector.metadata.deal_value ||
+            existingVector.metadata.conversion_days !== vector.metadata.conversion_days ||
+            existingVector.metadata.pipeline !== vector.metadata.pipeline ||
+            existingVector.metadata.dealstage !== vector.metadata.dealstage ||
+            existingVector.metadata.days_in_pipeline !== vector.metadata.days_in_pipeline;
+            
+          return dealMetadataChanged;
+        });
+        
+        // Only upsert if there are changes
+        if (vectorsToUpsert.length > 0) {
+          logger.info(`Upserting ${vectorsToUpsert.length} changed vectors`);
+          
+          // Prepare deal info for upserting
+          const dealInfo = {
+            deal_id: deal.id,
+            deal_value: parseFloat(deal.properties?.amount) || 0,
+            conversion_days: calculateConversionDays(deal.properties || {}),
+            pipeline: deal.properties?.pipeline || 'unknown',
+            dealstage: deal.properties?.dealstage || 'unknown',
+            days_in_pipeline: parseInt(deal.properties?.hs_time_in_pipeline) || 0
+          };
+          
+          // Call the proper upsert method
+          await pineconeClient.upsertVectorsWithDealMetadata(
+            namespace,
+            documentsWithEmbeddings.filter(doc => 
+              vectorsToUpsert.some(v => v.id === doc.metadata.id.toString())
+            ),
+            documentsWithEmbeddings
+              .filter(doc => vectorsToUpsert.some(v => v.id === doc.metadata.id.toString()))
+              .map(doc => ({ embedding: doc.embedding })),
+            dealInfo
+          );
+        } else {
+          logger.info(`No vectors changed, skipping upsert for deal ${deal.id}`);
+        }
+      } catch (fetchError) {
+        // If error fetching (like namespace doesn't exist), just upsert all
+        logger.warn(`Error checking existing vectors: ${fetchError.message}. Will upsert all vectors.`);
+        
+        // Upsert to Pinecone
+        const dealInfo = {
+          deal_id: deal.id,
+          deal_value: parseFloat(deal.properties?.amount) || 0,
+          conversion_days: calculateConversionDays(deal.properties || {}),
+          pipeline: deal.properties?.pipeline || 'unknown',
+          dealstage: deal.properties?.dealstage || 'unknown',
+          days_in_pipeline: parseInt(deal.properties?.hs_time_in_pipeline) || 0
+        };
+        
+        // Call the proper upsert method
+        await pineconeClient.upsertVectorsWithDealMetadata(
+          namespace,
+          documentsWithEmbeddings,
+          documentsWithEmbeddings.map(doc => ({ embedding: doc.embedding })),
+          dealInfo
+        );
+      }
+      
+    } catch (pineconeError) {
+      logger.error(`Error in Pinecone operations: ${pineconeError.message}`);
+      throw pineconeError;
+    }
+    
+    const processingDuration = (Date.now() - processingStart) / 1000;
+    logger.info(`Deal ${deal.id} processed in ${processingDuration.toFixed(2)} seconds`);
+    
+  } catch (error) {
+    const processingDuration = (Date.now() - processingStart) / 1000;
+    logger.error(`Deal ${deal.id} processing failed after ${processingDuration.toFixed(2)} seconds: ${error.message}`);
+    throw error;
+  }
+}
+
+/**
+ * Fetches contacts and companies associated with a deal
+ */
 async function getAssociatedRecords(hubspotClient: HubspotClient, deal: any) {
   const contacts: HubspotRecord[] = [];
   const companies: HubspotRecord[] = [];
 
   try {
-    logger.info(`Checking associations for deal ${deal.id}:`, {
-      hasAssociations: !!deal.associations,
-      contactsCount: deal.associations?.contacts?.results?.length || 0,
-      companiesCount: deal.associations?.companies?.results?.length || 0
-    });
-
-    // Get associated contacts sequentially
+    // If we don't have associations but have a deal ID, try to fetch associations directly
+    if ((!deal.associations || (!deal.associations?.contacts?.results?.length && !deal.associations?.companies?.results?.length)) && deal.id) {
+      try {
+        // Fetch associations directly using the HubSpot API
+        const associationsResult = await hubspotClient.getAssociations(deal.id, 'deal');
+        
+        if (associationsResult?.results) {
+          // Extract contact and company IDs
+          const contactIds = associationsResult.results.contacts?.map(a => a.id) || [];
+          const companyIds = associationsResult.results.companies?.map(a => a.id) || [];
+          
+          // Create a synthetic associations structure for processing below
+          if (!deal.associations) deal.associations = {};
+          if (contactIds.length > 0) {
+            deal.associations.contacts = {
+              results: contactIds.map(id => ({ id }))
+            };
+          }
+          if (companyIds.length > 0) {
+            deal.associations.companies = {
+              results: companyIds.map(id => ({ id }))
+            };
+          }
+        }
+      } catch (associationsError) {
+        logger.error(`Error fetching associations: ${associationsError.message}`);
+      }
+    }
+    
+    // Check for associated contacts
     if (deal.associations?.contacts?.results?.length > 0) {
-      logger.info(`Fetching ${deal.associations.contacts.results.length} associated contacts for deal ${deal.id}`);
-      
       for (const contact of deal.associations.contacts.results) {
         try {
           const contactData = await hubspotClient.getContact(contact.id);
           if (contactData) {
-            logger.info(`Successfully fetched contact ${contact.id} for deal ${deal.id}`);
             contacts.push(contactData);
           }
-          // Add delay between contact fetches
-          await sleep(1000);
+          await sleep(1000); // Rate limiting
         } catch (error) {
-          logger.error(`Error fetching contact ${contact.id}:`, error);
+          logger.error(`Error fetching contact ${contact.id}: ${error.message}`);
         }
       }
-    } else {
-      logger.info(`No contacts associated with deal ${deal.id}`);
     }
 
-    // Get associated companies sequentially
+    // Check for associated companies
     if (deal.associations?.companies?.results?.length > 0) {
-      logger.info(`Fetching ${deal.associations.companies.results.length} associated companies for deal ${deal.id}`);
-      
       for (const company of deal.associations.companies.results) {
         try {
           const companyData = await hubspotClient.getCompany(company.id);
           if (companyData) {
-            logger.info(`Successfully fetched company ${company.id} for deal ${deal.id}`);
             companies.push(companyData);
           }
-          // Add delay between company fetches
-          await sleep(1000);
+          await sleep(1000); // Rate limiting
         } catch (error) {
-          logger.error(`Error fetching company ${company.id}:`, error);
+          logger.error(`Error fetching company ${company.id}: ${error.message}`);
         }
       }
-    } else {
-      logger.info(`No companies associated with deal ${deal.id}`);
     }
 
-    logger.info(`Successfully fetched ${contacts.length} contacts and ${companies.length} companies for deal ${deal.id}`);
     return { contacts, companies };
   } catch (error) {
-    logger.error(`Error fetching associated records for deal ${deal.id}:`, error);
-    // Return what we have so far rather than failing completely
+    logger.error(`Error fetching associated records: ${error.message}`);
+    // Return whatever we managed to retrieve
     return { contacts, companies };
   }
 }
 
-function calculateDealMetrics(deal: any, contacts: any[], companies: any[]): DealMetadata {
-  const createDate = new Date(deal.properties.createdate);
-  const closeDate = deal.properties.closedate ? new Date(deal.properties.closedate) : null;
-  const daysInPipeline = deal.properties.hs_time_in_pipeline ? 
-    parseInt(deal.properties.hs_time_in_pipeline) : 
-    (closeDate ? Math.ceil((closeDate.getTime() - createDate.getTime()) / (1000 * 60 * 60 * 24)) : 0);
-
-  return {
-    deal_id: deal.id,
-    deal_value: parseFloat(deal.properties.amount) || 0,
-    conversion_days: calculateConversionDays(deal.properties),
-    pipeline: deal.properties.pipeline || 'default',
-    dealstage: deal.properties.dealstage,
-    won_date: deal.properties.hs_date_entered_closedwon || '',
-    lost_date: deal.properties.hs_date_entered_closedlost || '',
-    related_contacts: contacts.map(c => c.id),
-    related_companies: companies.map(c => c.id),
-    days_in_pipeline: daysInPipeline,
-    stage_changes: parseInt(deal.properties.hs_deal_stage_changes) || 0
-  };
-}
-
-function calculateContactMetrics(contact: any, dealHistory: any[]): ContactMetadata {
-  const wonDeals = dealHistory.filter(d => d.properties.dealstage === 'closedwon').length;
-  const lostDeals = dealHistory.filter(d => d.properties.dealstage === 'closedlost').length;
-  
-  return {
-    contact_id: contact.id,
-    lifecycle_stage: contact.properties.lifecyclestage || 'unknown',
-    lead_source: contact.properties.lead_source || 'unknown',
-    total_deals_participated: dealHistory.length,
-    total_won_deals: wonDeals,
-    total_lost_deals: lostDeals,
-    first_conversion_date: contact.properties.first_conversion_date || '',
-    last_activity_date: contact.properties.last_activity_date || '',
-    engagement_score: parseFloat(contact.properties.hs_email_engagement) || 0,
-    deal_participation_rate: dealHistory.length > 0 ? wonDeals / dealHistory.length : 0,
-    industry: contact.properties.industry || '',
-    company_size: contact.properties.numemployees || ''
-  };
-}
-
-function calculateCompanyMetrics(company: any): CompanyMetadata {
-  return {
-    company_id: company.id,
-    industry: company.properties.industry || '',
-    company_size: company.properties.numberofemployees || '',
-    total_revenue: parseFloat(company.properties.total_revenue) || 0,
-    country: company.properties.country || '',
-    city: company.properties.city || '',
-    total_deals: parseInt(company.properties.total_deals) || 0,
-    won_deals: parseInt(company.properties.closed_won_deals) || 0,
-    lost_deals: parseInt(company.properties.closed_lost_deals) || 0,
-    average_deal_size: parseFloat(company.properties.avg_deal_value) || 0,
-    average_sales_cycle: parseFloat(company.properties.avg_sales_cycle) || 0,
-    engagement_score: parseFloat(company.properties.engagement_score) || 0,
-    first_deal_date: company.properties.first_deal_created_date || '',
-    last_deal_date: company.properties.recent_deal_close_date || ''
-  };
-}
-
-// Add this new helper function near the other metadata calculation functions
-function enrichMetadataWithDealInfo(baseMetadata: any, dealMetadata: DealMetadata | null = null): any {
-  if (!dealMetadata) {
-    return baseMetadata;
-  }
-
-  return {
-    ...baseMetadata,
-    deal_id: dealMetadata.deal_id,
-    deal_value: dealMetadata.deal_value,
-    conversion_days: dealMetadata.conversion_days,
-    pipeline: dealMetadata.pipeline,
-    dealstage: dealMetadata.dealstage,
-    days_in_pipeline: dealMetadata.days_in_pipeline
-  };
-}
-
-// Modify the createDocuments function to use the new helper
-async function createDocuments(documentPackager: DocumentPackager, records: any[], type: 'deal' | 'contact' | 'company', portalId: string, classification: 'ideal' | 'nonideal', dealMetadata: any) {
-  return Promise.all(
-    records.map(async (record) => {
-      const doc = await documentPackager.packageDocument(record, type, portalId);
-      
-      // Calculate specific metadata based on record type
-      let specificMetadata = {};
-      if (type === 'deal') {
-        specificMetadata = calculateDealMetrics(record, [], []); // Empty arrays since we don't need recursive metadata
-      } else if (type === 'contact') {
-        specificMetadata = calculateContactMetrics(record, [dealMetadata]); // Pass the parent deal
-      } else if (type === 'company') {
-        specificMetadata = calculateCompanyMetrics(record);
-      }
-
-      // Enrich the metadata with deal information
-      const enrichedMetadata = enrichMetadataWithDealInfo(
-        { ...doc.metadata, ...specificMetadata },
-        type === 'deal' ? specificMetadata : dealMetadata
-      );
-
-      return {
-        ...doc,
-        metadata: {
-          ...enrichedMetadata,
-          classification,
-          source: 'auto-train',
-          record_type: type
-        }
-      };
-    })
-  );
-}
-
-async function refreshHubSpotToken(refreshToken: string): Promise<{ access_token: string; refresh_token: string; expires_in: number }> {
-  const response = await fetch('https://api.hubapi.com/oauth/v1/token', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    body: new URLSearchParams({
-      grant_type: 'refresh_token',
-      client_id: Deno.env.get('HUBSPOT_CLIENT_ID')!,
-      client_secret: Deno.env.get('HUBSPOT_CLIENT_SECRET')!,
-      refresh_token: refreshToken,
-    }).toString(),
-  });
-
-  if (!response.ok) {
-    const error = await response.text();
-    logger.error('Failed to refresh token:', error);
-    throw new Error('Failed to refresh HubSpot token');
-  }
-
-  return response.json();
-}
-
-async function validateAndRefreshTokenIfNeeded(
-  hubspotClient: HubspotClient, 
-  portalId: string, 
-  currentToken: string,
-  refreshToken: string
-): Promise<string> {
-  try {
-    logger.info(`[Token Validation] Starting token validation for portal ${portalId}`);
-    
-    // Check if we need to refresh the token
-    try {
-      const newTokens = await refreshHubSpotToken(refreshToken);
-      logger.info('[Token Validation] Successfully obtained new tokens');
-      
-      // Update tokens in database
-      logger.info('[Token Validation] Updating tokens in database');
-      const supabase = createClient(
-        Deno.env.get('SUPABASE_URL')!,
-        Deno.env.get('SUPABASE_SERVICE_ROLE_KEY')!
-      );
-
-      const newEncryptedToken = await encrypt(newTokens.access_token, Deno.env.get('ENCRYPTION_KEY')!);
-      const newEncryptedRefreshToken = await encrypt(newTokens.refresh_token, Deno.env.get('ENCRYPTION_KEY')!);
-      
-      const { error: updateError } = await supabase
-        .from('hubspot_accounts')
-        .update({
-          access_token: newEncryptedToken,
-          refresh_token: newEncryptedRefreshToken,
-          expires_at: new Date(Date.now() + newTokens.expires_in * 1000).toISOString(),
-          updated_at: new Date().toISOString()
-        })
-        .eq('portal_id', portalId);
-        
-      if (updateError) {
-        logger.error('[Token Validation] Failed to update tokens in database:', updateError);
-        throw new Error('Failed to update HubSpot tokens');
-      }
-      
-      // Update the client with the new token
-      hubspotClient.updateToken(newTokens.access_token);
-      logger.info('[Token Validation] Successfully updated tokens in database');
-      return newTokens.access_token;
-      
-    } catch (refreshError) {
-      logger.error('[Token Validation] Error during token refresh:', {
-        message: refreshError.message,
-        stack: refreshError.stack,
-        details: refreshError
-      });
-      throw refreshError;
-    }
-  } catch (error) {
-    logger.error(`[Token Validation] Fatal error for portal ${portalId}:`, {
-      message: error.message,
-      stack: error.stack,
-      details: error
-    });
-    throw error;
-  }
-}
-
-// Add a new function to handle token refresh during API calls
-async function handleApiCall<T>(
-  hubspotClient: HubspotClient,
-  portalId: string,
-  refreshToken: string,
-  apiCall: () => Promise<T>
-): Promise<T> {
-  try {
-    return await apiCall();
-  } catch (error) {
-    if (error.message?.includes('401') || error.message?.includes('expired')) {
-      logger.info(`Token expired during API call for portal ${portalId}, refreshing...`);
-      const newTokens = await refreshHubSpotToken(refreshToken);
-      hubspotClient.updateToken(newTokens.access_token);
-      logger.info('Token refreshed, retrying API call');
-      return await apiCall();
-    }
-    throw error;
-  }
-}
-
-async function processDeals(
-  deals: any[], 
-  classification: 'ideal' | 'nonideal',
-  hubspotClient: HubspotClient,
-  documentPackager: DocumentPackager,
-  openai: OpenAI,
+/**
+ * Get existing vectors for a list of deal IDs
+ */
+async function getExistingVectors(
   pineconeClient: PineconeClient,
-  portalId: string
-) {
-  logger.info(`Starting to process ${deals.length} ${classification} deals`);
-  
-  // Process deals in smaller batches
-  const dealBatches = chunkArray(deals, 5);
-  logger.info(`Split deals into ${dealBatches.length} batches of 5`);
-  
-  for (const batch of dealBatches) {
-    logger.info(`Starting batch processing of ${batch.length} ${classification} deals`);
-    logger.info('Batch deal IDs:', batch.map(d => d.id));
-    
-    for (const deal of batch) {
-      try {
-        logger.info(`\n=== Starting processing of deal ${deal.id} ===`);
-        logger.info('Deal details:', {
-          id: deal.id,
-          name: deal.properties?.dealname,
-          stage: deal.properties?.dealstage,
-          hasAssociations: !!deal.associations
-        });
-
-        await sleep(2000);
-        logger.info('Initial delay complete');
-
-        // Get associated records
-        logger.info(`Starting to fetch associated records for deal ${deal.id}`);
-        const { contacts, companies } = await getAssociatedRecords(hubspotClient, deal);
-        logger.info(`Completed fetching associated records: ${contacts.length} contacts, ${companies.length} companies`);
-
-        // Calculate deal metrics
-        logger.info(`Starting to calculate metrics for deal ${deal.id}`);
-        const dealMetadata = calculateDealMetrics(deal, contacts, companies);
-        logger.info('Deal metrics calculated:', dealMetadata);
-
-        // Create documents
-        logger.info(`Starting document creation for deal ${deal.id}`);
-        
-        logger.info('Creating deal documents...');
-        const dealDocs = await createDocuments(documentPackager, [deal], 'deal', portalId, classification, dealMetadata);
-        logger.info(`Created ${dealDocs.length} deal documents`);
-        await sleep(2000);
-        
-        logger.info('Creating contact documents...');
-        const contactDocs = contacts.length > 0 ? 
-          await createDocuments(documentPackager, contacts, 'contact', portalId, classification, dealMetadata) : 
-          [];
-        logger.info(`Created ${contactDocs.length} contact documents`);
-        await sleep(2000);
-        
-        logger.info('Creating company documents...');
-        const companyDocs = companies.length > 0 ? 
-          await createDocuments(documentPackager, companies, 'company', portalId, classification, dealMetadata) : 
-          [];
-        logger.info(`Created ${companyDocs.length} company documents`);
-
-        const documents = [...dealDocs, ...contactDocs, ...companyDocs];
-        logger.info(`Total documents created: ${documents.length}`);
-        logger.info('Document metadata IDs:', documents.map(doc => doc.metadata.id));
-
-        // Get embeddings
-        logger.info(`Starting OpenAI embedding creation for ${documents.length} documents`);
-        const embeddingResponse = await openai.embeddings.create({
-          model: 'text-embedding-3-large',
-          input: documents.map(doc => doc.content)
-        });
-        logger.info(`Successfully received ${embeddingResponse.data.length} embeddings from OpenAI`);
-
-        await sleep(2000);
-        logger.info('Post-embedding delay complete');
-
-        // Create deal info object for metadata consistency
-        const dealInfo = {
-          deal_id: deal.id,
-          deal_value: parseFloat(deal.properties?.amount) || 0,
-          conversion_days: calculateConversionDays(deal.properties),
-          pipeline: deal.properties?.pipeline || 'unknown',
-          dealstage: deal.properties?.dealstage || 'unknown',
-          days_in_pipeline: parseInt(deal.properties?.hs_time_in_pipeline) || 0
-        };
-
-        // Use the Pinecone client to upsert vectors with deal metadata
-        const namespace = `hubspot-${portalId}`;
-        try {
-          logger.info(`Deleting namespace ${namespace} if it exists`);
-          await pineconeClient.deleteNamespace(namespace);
-          logger.info(`Successfully deleted namespace ${namespace}`);
-          
-          logger.info(`Upserting vectors to namespace ${namespace} with deal ID ${dealInfo.deal_id}`);
-          const upsertResult = await pineconeClient.upsertVectorsWithDealMetadata(
-            namespace,
-            documents,
-            embeddingResponse.data,
-            dealInfo
-          );
-          logger.info('Pinecone upsert completed:', upsertResult);
-        } catch (pineconeError) {
-          logger.error('Pinecone operation failed:', {
-            error: pineconeError.message,
-            stack: pineconeError.stack,
-            namespace,
-            dealId: deal.id
-          });
-          throw pineconeError;
-        }
-
-        logger.info(`=== Successfully completed processing deal ${deal.id} ===\n`);
-        await sleep(3000);
-        
-      } catch (error) {
-        logger.error(`Error processing deal ${deal.id}:`, {
-          error: error.message,
-          stack: error.stack,
-          dealId: deal.id,
-          dealName: deal.properties?.dealname
-        });
-        await sleep(5000);
-      }
-    }
-    
-    logger.info(`\n=== Completed processing batch of ${batch.length} deals ===`);
-    await sleep(10000);
-  }
-  
-  logger.info(`\n=== Completed processing all ${classification} deals ===`);
-}
-
-async function processRecords(
-  records: any[], 
-  type: string, 
-  portalId: string, 
-  hubspotClient: HubspotClient,
-  refreshToken: string,
-  shouldClearNamespace: boolean = false
-) {
-  const logger = new Logger('processRecords');
-  logger.info(`Processing ${records.length} ${type} records for portal ${portalId}`);
-
-  // Query HubSpot for ideal and less ideal counts using handleApiCall
-  const idealQuery = await handleApiCall(hubspotClient, portalId, refreshToken, () =>
-    hubspotClient.searchRecords(type, {
-      filterGroups: [{
-        filters: [{
-          propertyName: 'training_score',
-          operator: 'GTE',
-          value: '80'
-        }]
-      }],
-      limit: 1,
-      after: 0
-    })
-  );
-
-  const lessIdealQuery = await handleApiCall(hubspotClient, portalId, refreshToken, () =>
-    hubspotClient.searchRecords(type, {
-      filterGroups: [{
-        filters: [{
-          propertyName: 'training_score',
-          operator: 'LT',
-          value: '50'
-        }]
-      }],
-      limit: 1,
-      after: 0
-    })
-  );
-
-  const currentIdealCount = idealQuery.total;
-  const currentLessIdealCount = lessIdealQuery.total;
-
-  // Update counts in database based on record type
-  const supabase = createClient(
-    Deno.env.get('SUPABASE_URL')!,
-    Deno.env.get('SUPABASE_SERVICE_ROLE_KEY')!
-  );
-
-  const updateData = {
-    last_training_date: new Date().toISOString()
-  };
-
-  // Add the appropriate counts based on record type
-  if (type === 'contacts') {
-    Object.assign(updateData, {
-      current_ideal_contacts: currentIdealCount,
-      current_less_ideal_contacts: currentLessIdealCount
-    });
-  } else if (type === 'companies') {
-    Object.assign(updateData, {
-      current_ideal_companies: currentIdealCount,
-      current_less_ideal_companies: currentLessIdealCount
-    });
-  } else if (type === 'deals') {
-    Object.assign(updateData, {
-      current_ideal_deals: currentIdealCount,
-      current_less_ideal_deals: currentLessIdealCount
-    });
-  }
-
-  const { error: updateError } = await supabase
-    .from('hubspot_accounts')
-    .update(updateData)
-    .eq('portal_id', portalId);
-
-  if (updateError) {
-    logger.error('Failed to update training metrics:', updateError);
-    throw new Error('Failed to update training metrics');
-  }
-
-  // Initialize OpenAI client
-  const openai = new OpenAI({
-    apiKey: Deno.env.get('OPENAI_API_KEY')!
-  });
-
-  // Initialize Pinecone client
-  const pineconeClient = new PineconeClient();
-  await pineconeClient.initialize(Deno.env.get('PINECONE_API_KEY')!, Deno.env.get('PINECONE_INDEX')!);
-  
-  const namespace = `hubspot-${portalId}`;
-
-  // Delete all existing vectors in the namespace if this is the first record type
-  if (shouldClearNamespace) {
-    try {
-      logger.info(`Attempting to delete existing vectors in namespace: ${namespace}`);
-      await pineconeClient.deleteNamespace(namespace);
-      logger.info(`Successfully deleted vectors in namespace: ${namespace}`);
-    } catch (error) {
-      // If the namespace doesn't exist (404), we can ignore the error
-      if (error.message.includes('404')) {
-        logger.info(`Namespace ${namespace} does not exist, skipping deletion`);
-      } else {
-        // For other errors, log them but continue processing
-        logger.warn(`Error deleting vectors in namespace ${namespace}:`, error);
-      }
-    }
-  }
-
-  const documentPackager = new DocumentPackager(hubspotClient);
-
-  // Process each record through the document packager
-  const documents = await Promise.all(
-    records.map(record => documentPackager.packageDocument(record, type as 'contact' | 'company' | 'deal', portalId))
-  );
-
-  logger.info(`Created ${documents.length} documents`);
-
-  // Get embeddings for all documents using OpenAI
-  const embeddingResponse = await openai.embeddings.create({
-    model: 'text-embedding-3-large',
-    input: documents.map(doc => doc.content)
-  });
-
-  logger.info('Embedding response structure:', {
-    hasData: !!embeddingResponse.data,
-    firstEmbeddingLength: embeddingResponse.data[0]?.embedding?.length,
-    isEmbeddingArray: Array.isArray(embeddingResponse.data[0]?.embedding)
-  });
-
-  // Create vectors for all documents
-  const vectors = documents.map((doc, index) => ({
-    id: doc.metadata.id.toString(),
-    values: Array.from(embeddingResponse.data[index].embedding),
-    metadata: {
-      id: doc.metadata.id.toString(),
-      recordType: doc.metadata.recordType,
-      updatedAt: doc.metadata.updatedAt
-    }
-  }));
-
-  logger.info('Vector structure:', {
-    totalVectors: vectors.length,
-    firstVector: {
-      hasId: !!vectors[0]?.id,
-      valuesLength: vectors[0]?.values?.length,
-      isValuesArray: Array.isArray(vectors[0]?.values),
-      metadata: vectors[0]?.metadata
-    },
-    namespace
-  });
-
+  dealIds: string[],
+  namespace: string
+): Promise<any[]> {
   try {
-    logger.info(`Attempting to upsert ${vectors.length} vectors into namespace: ${namespace}`);
-    await pineconeClient.upsertVectorsWithDealMetadata(
-      namespace,
-      documents,
-      embeddingResponse.data,
-      { 
-        deal_id: '', 
-        deal_value: 0, 
-        conversion_days: 0, 
-        pipeline: '', 
-        dealstage: '',
-        days_in_pipeline: 0
+    logger.info(`Fetching existing vectors for ${dealIds.length} deals from namespace ${namespace}`);
+    logger.info(`Deal IDs to check: ${dealIds.join(', ')}`);
+    
+    if (!dealIds.length) {
+      logger.info('No deal IDs provided, returning empty array');
+      return [];
+    }
+    
+    // Get index host and API key from environment
+    const pineconeIndexHost = Deno.env.get('PINECONE_INDEX_HOST');
+    const pineconeApiKey = Deno.env.get('PINECONE_API_KEY');
+    
+    if (!pineconeIndexHost || !pineconeApiKey) {
+      logger.error('Missing PINECONE_INDEX_HOST or PINECONE_API_KEY environment variables');
+      return [];
+    }
+    
+    // Use direct fetchByIds method which is more reliable than query
+    try {
+      logger.info(`Trying direct fetch by IDs for ${dealIds.length} deals in namespace: ${namespace}`);
+      const fetchResult = await pineconeClient.fetchByIds(
+        dealIds.map(id => id.toString()), 
+        namespace,
+        pineconeIndexHost,
+        pineconeApiKey
+      );
+      
+      logger.info(`Direct fetch result structure: ${JSON.stringify(Object.keys(fetchResult))}`);
+      
+      // Check for matches property (pineconeClient.fetchByIds returns matches, not vectors)
+      if (fetchResult.matches && fetchResult.matches.length > 0) {
+        logger.info(`Found ${fetchResult.matches.length} vectors with IDs: ${fetchResult.matches.map(m => m.id).join(', ')}`);
+        
+        // Log more details about what we found
+        if (fetchResult.matches[0].metadata) {
+          logger.info(`Vector sample metadata keys: ${Object.keys(fetchResult.matches[0].metadata || {}).join(', ')}`);
+        }
+        
+        // Return the matches directly since they're already in the format we need
+        logger.info(`Successfully found ${fetchResult.matches.length} vectors from result`);
+        return fetchResult.matches;
       }
-    );
-    logger.info(`Successfully upserted ${vectors.length} vectors into namespace: ${namespace}`);
+      
+      logger.info('No matching vectors found with direct fetch');
+      return [];
+    } catch (fetchError) {
+      logger.error(`Error fetching vectors by ID: ${fetchError.message}`);
+      return [];
+    }
   } catch (error) {
-    logger.error('Error upserting vectors:', error);
-    logger.error('Error upserting vectors - first vector:', JSON.stringify(vectors[0]));
-    logger.error('Attempted namespace:', namespace);
-    throw error;
+    logger.error(`Error getting existing vectors: ${error.message}`);
+    return [];
   }
-
-  return vectors.length; // Return the number of vectors processed
 }
 
+/**
+ * Main handler for the edge function
+ */
 serve(async (req) => {
   try {
-    logger.info('Starting process training function');
+    logger.info('Starting hubspot-auto-training-v2');
     
     // Handle CORS preflight
     if (req.method === "OPTIONS") {
@@ -930,21 +837,21 @@ serve(async (req) => {
     // Fetch all HubSpot accounts
     const { data: accounts, error: accountsError } = await supabase
       .from('hubspot_accounts')
-      .select('*');
+      .select('*')
+      .eq('status', 'active');
 
     if (accountsError) {
-      logger.error('Error fetching accounts from Supabase:', accountsError);
       throw new Error(`Failed to fetch accounts: ${accountsError.message}`);
     }
 
     if (!accounts || accounts.length === 0) {
-      throw new Error('No HubSpot accounts found');
+      throw new Error('No active HubSpot accounts found');
     }
 
-    logger.info(`Found ${accounts.length} HubSpot accounts to process`);
+    logger.info(`Found ${accounts.length} active HubSpot accounts to process`);
 
-    // Process each account
-    const allResults = {
+    // Track results
+    const results = {
       total: accounts.length,
       processed: 0,
       successful: 0,
@@ -952,97 +859,101 @@ serve(async (req) => {
       portals: [] as any[]
     };
 
+    // Process each account
     for (const account of accounts) {
       const portalId = account.portal_id;
       logger.info(`Processing portal ${portalId}`);
 
       try {
-        // Decrypt tokens
-        let decryptedToken = await decrypt(account.access_token, Deno.env.get('ENCRYPTION_KEY')!);
+        // Initialize clients and services
+        const decryptedToken = await decrypt(account.access_token, Deno.env.get('ENCRYPTION_KEY')!);
         const decryptedRefreshToken = await decrypt(account.refresh_token, Deno.env.get('ENCRYPTION_KEY')!);
         
         if (!decryptedToken || !decryptedRefreshToken) {
-          throw new Error('HubSpot tokens are missing or invalid');
+          throw new Error('Invalid HubSpot tokens');
         }
+        
+        logger.info(`Initializing services for portal ${portalId}`);
 
-        // Initialize HubSpot client with current token
         const hubspotClient = new HubspotClient(decryptedToken);
-
-        // Test token and refresh if needed
-        try {
-          await handleApiCall(hubspotClient, portalId, decryptedRefreshToken, () => 
-            hubspotClient.searchRecords('contacts', { limit: 1 })
-          );
-          logger.info('Token validation successful');
-        } catch (error) {
-          logger.error(`Token validation failed for portal ${portalId}:`, error);
-          throw error;
-        }
-
-        // Process each record type
-        const recordTypes = ['contacts', 'companies', 'deals'];
-        const portalResults = {
+        const openaiApiKey = Deno.env.get('OPENAI_API_KEY')!;
+        const openai = new OpenAI({ apiKey: openaiApiKey });
+        const pineconeApiKey = Deno.env.get('PINECONE_API_KEY')!;
+        const pineconeIndex = Deno.env.get('PINECONE_INDEX')!;
+        const pineconeClient = new PineconeClient();
+        await pineconeClient.initialize(pineconeApiKey, pineconeIndex);
+        const documentPackager = new DocumentPackager(hubspotClient);
+        logger.info(`Services initialized for portal ${portalId}`);
+        
+        // Validate token
+        logger.info(`Validating token for portal ${portalId}`);
+        await handleApiCall(hubspotClient, portalId, decryptedRefreshToken, () => 
+          hubspotClient.searchRecords('contacts', { limit: 1 })
+        );
+        
+        // Track portal results
+        const portalResult = {
           portalId,
-          contacts: { processed: 0, successful: 0, failed: 0, errors: [] as any[] },
-          companies: { processed: 0, successful: 0, failed: 0, errors: [] as any[] },
-          deals: { processed: 0, successful: 0, failed: 0, errors: [] as any[] }
+          ideal: { total: 0, processed: 0 },
+          nonideal: { total: 0, processed: 0 }
         };
-
-        for (const type of recordTypes) {
-          try {
-            logger.info(`Processing ${type} for portal ${portalId}`);
-            
-            // Use handleApiCall for the search
-            const searchResponse = await handleApiCall(hubspotClient, portalId, decryptedRefreshToken, () =>
-              hubspotClient.searchRecords(type, {
-                limit: 100
-              })
-            );
-
-            if (searchResponse.total > 0) {
-              portalResults[type].processed = searchResponse.total;
-              await processRecords(
-                searchResponse.results, 
-                type, 
-                portalId.toString(), 
-                hubspotClient,
-                decryptedRefreshToken,
-                type === 'contacts'
-              );
-              portalResults[type].successful = searchResponse.total;
-            }
-          } catch (error) {
-            portalResults[type].failed = 1;
-            const errorDetails = {
-              message: error.message,
-              stack: error.stack,
-              cause: error.cause,
-              type: error.name
-            };
-            portalResults[type].errors.push(errorDetails);
-            logger.error(`Error processing ${type} for portal ${portalId}:`, errorDetails);
-          }
+        
+        // Process ideal deals
+        try {
+          logger.info(`Starting ideal deals processing for portal ${portalId}`);
+          
+          const idealDeals = await getDealsForTraining(hubspotClient, 'ideal', portalId, decryptedRefreshToken, documentPackager, openai, pineconeClient);
+          
+          portalResult.ideal.total = idealDeals.total;
+          portalResult.ideal.processed = idealDeals.processed;
+          
+          logger.info(`Processed ${idealDeals.processed} ideal deals for portal ${portalId}`);
+          
+        } catch (idealDealsError) {
+          logger.error(`Error processing ideal deals for portal ${portalId}: ${idealDealsError.message}`);
         }
-
-        allResults.portals.push(portalResults);
-        allResults.successful++;
+        
+        // Process non-ideal deals
+        logger.info(`Starting non-ideal deals processing for portal ${portalId}`);
+        const nonIdealDeals = await getDealsForTraining(hubspotClient, 'nonideal', portalId, decryptedRefreshToken, documentPackager, openai, pineconeClient);
+        portalResult.nonideal.total = nonIdealDeals.total;
+        portalResult.nonideal.processed = nonIdealDeals.processed;
+        
+        logger.info(`Processed ${nonIdealDeals.processed} non-ideal deals for portal ${portalId}`);
+        
+        // Update database with metrics
+        await supabase
+          .from('hubspot_accounts')
+          .update({
+            last_training_date: new Date().toISOString(),
+            current_ideal_deals: portalResult.ideal.processed,
+            current_less_ideal_deals: portalResult.nonideal.processed
+          })
+          .eq('portal_id', portalId);
+        
+        // Track success
+        results.portals.push(portalResult);
+        results.successful++;
+        logger.info(`Successfully processed portal ${portalId}`);
       } catch (error) {
-        allResults.failed++;
-        logger.error(`Failed to process portal ${portalId}:`, error);
-        allResults.portals.push({
+        // Track failure
+        results.failed++;
+        logger.error(`Failed to process portal ${portalId}: ${error.message}`);
+        results.portals.push({
           portalId,
           error: error.message
         });
       }
-
-      allResults.processed++;
+      
+      results.processed++;
     }
 
+    // Return the response
     return new Response(
       JSON.stringify({
         success: true,
-        message: `Processed ${allResults.processed} portals`,
-        results: allResults
+        message: `Processed ${results.processed} HubSpot accounts`,
+        results
       }),
       {
         headers: { "Content-Type": "application/json" },
@@ -1050,7 +961,7 @@ serve(async (req) => {
       }
     );
   } catch (error) {
-    logger.error('Error in process training function:', error);
+    logger.error('Error in hubspot-auto-training-v2:', error);
     return new Response(
       JSON.stringify({
         success: false,
@@ -1062,4 +973,4 @@ serve(async (req) => {
       }
     );
   }
-});
+});
